Remove unused code snippets
