+++ src/components/registrationForm/RegistrationForm.tsx
@@ -2,7 +2,7 @@
  * @module RegistrationForm
  * @description Module containing the component for rendering the registration form.
  */
-import { UserContext } from '@contexts/AuthContext';
+import { AuthContext } from '@contexts/AuthContext';
 import { RegistrationData } from '@data/formData';
 import { newUserSchema } from '@data/formSchemas';
 import { zodResolver } from '@hookform/resolvers/zod';
