+++ src/main.tsx
@@ -9,6 +9,7 @@
  */
 
 import { BackdropProvider } from '@contexts/BackdropContext.tsx';
+
 import router from '@router/Router.tsx';
 import * as React from 'react';
 import * as ReactDOM from 'react-dom/client';
