+++ src/middleware/authHandler.ts
@@ -1,4 +1,4 @@
-import { UserContext } from '@contexts/AuthContext';
+import { AuthContext } from '@contexts/AuthContext';
 import { useContext, useEffect, useState } from 'react';
 import { NavigateFunction } from 'react-router-dom';
 import { getUserRole } from './apiHandler';
